refactor: move drop related services from waifu.ts to fxart.ts feat(fxart.ts): add new service file for fxart related functions fix: update import paths to use fxart.ts instead of waifu.ts feat: add holderOfCollection property in GenerativeLayout.vue refactor(waifu.ts): remove drop related functions and update baseURL
