+++ services/waifu.ts
@@ -1,46 +1,13 @@
 import { $fetch, FetchError } from 'ofetch'
-import type { DropItem } from '@/params/types'
+const WAIFU_BASE_URL = 'https://waifu-me.kodadot.workers.dev'
 
 const api = $fetch.create({
-  baseURL: 'https://fxart.kodadot.workers.dev/',
+  baseURL: WAIFU_BASE_URL,
 })
 
 // type ClaimResponse = Response<any>
 type Response<T> = {
   result: T
-}
-
-export const getDrops = async () => {
-  return await api<DropItem[]>('drops', {
-    method: 'GET',
-  })
-}
-
-export const getDropById = async (id: string) => {
-  return await api<DropItem>(`/drops/${id}`, {
-    method: 'GET',
-  })
-}
-
-export const getDropStatus = async (alias: string) => {
-  return await api<{ count: number }>(`/drops/${alias}/status`, {
-    method: 'GET',
-  })
-}
-
-export type DropMintedStatus = {
-  created_at: string
-  id: number
-  image: string
-  metadata: string
-  claimed: number
-  email: string
-  hash: string
-}
-export const getDropMintedStatus = async (alias: string, accountId: string) => {
-  return await api<DropMintedStatus>(`/drops/${alias}/accounts/${accountId}`, {
-    method: 'GET',
-  })
 }
 
 export const getLatestWaifuImages = async () => {
@@ -131,34 +98,6 @@ export const claimDropItem = async (
   })
 
   return value
-}
-
-export const allocateCollection = async (body, id) => {
-  try {
-    const response = await api(`/drops/allocate/${id}`, {
-      method: 'POST',
-      body,
-    })
-
-    return response
-  } catch (error) {
-    throw new Error(`[WAIFU::ALLOCATE] ERROR: ${(error as FetchError).data}`)
-  }
-}
-
-export const allocateClaim = async (body, id) => {
-  try {
-    const response = await api(`/drops/do/${id}`, {
-      method: 'post',
-      body,
-    })
-
-    return response
-  } catch (error) {
-    throw new Error(
-      `[WAIFU::ALLOCATE::CLAIM] ERROR: ${(error as FetchError).data}`,
-    )
-  }
 }
 
 export default api
